requirements.txt
absl-py==1.3.0
gin-config==0.5.0
numpy==1.23.5
qiskit==0.39.4
sympy==1.11.1
typing==3.7.4.3
z3-solver==4.11.2.0